+++ src/tiled/mainwindow.h
@@ -274,7 +274,7 @@ private:
     QPointer<PreferencesDialog> mPreferencesDialog;
 
     QMap<QMainWindow*, QByteArray> mMainWindowStates;
-    bool mHasRestoredLayout = false;
+    bool mReapplyLayoutOnResize = false;
 
     SessionOption<QStringList> mLoadedWorlds { "loadedWorlds" };