+++ NEWS.md
@@ -7,11 +7,14 @@
 * Added an action to create a world containing the current map (by Kanishka, #4562)
 * Made switching to the previously selected tool when pressing its shortcut again optional and off by default (by dogboydog, #4540)
 * Persisted collapsed state of the properties groups in the session (#4561)
+* Scripting: Added a script server and --eval, --eval-file and --script-server options (#4597)
 * Scripting: Added 'tiled.cell' function, 'cell.flags' property and 'TileLayerEdit.setCell' function (#4538)
 * Scripting: Added MapObject.resolvedClassName() (by MatusGuy, #4529)
-* Scripting: Added a script server and --eval, --eval-file and --script-server options, for running scripts in a running Tiled instance from the command-line
+* Scripting: Added Tileset.selectedTilesChanged signal (#4615)
+* Scripting: Fixed world scripting APIs to work without the editor (by Shubham Padkonde, #4584)
 * Fixed crash when the selection becomes empty while starting a move (#4536)
 * Fixed Properties view update on 'Reset Template Instance' and 'Replace With Template' actions
+* Fixed restoring of the layout for maximized windows on startup (#4580)
 * Linux: Added file associations for .tmj, .tsj, .tiled-project and .world files (by miffe, #4550)
 * Linux: Fixed the window icon on some Wayland compositors (by Balló György, #4567)
 * macOS: Declared UTIs and file associations for all supported Tiled formats (with Jyotish, #4469)